+++ udev.py
@@ -285,6 +285,24 @@ class Enumerator(object):
                 yield device
 
 
+class NoSuchDeviceError(LookupError):
+    """
+    An error indicating that no :class:`Device` was found for a specific
+    path.
+    """
+
+    def __init__(self, sys_path):
+        LookupError.__init__(self, sys_path)
+
+    @property
+    def sys_path(self):
+        """The path that caused this error"""
+        return self.args[0]
+
+    def __str__(self):
+        return 'No such device: {0!r}'.format(self.sys_path)
+
+
 class Device(Mapping):
     """
     A single device with attached attributes and properties.
@@ -323,15 +341,19 @@ class Device(Mapping):
         ``sys_path`` is a unicode or byte string containing the path of the
         device inside ``sysfs`` with the mount point included.
 
-        Return a :class:`Device` object for the device, or ``None``, if no
-        device existed at ``sys_path``.
+        Return a :class:`Device` object for the device.  Raise
+        :exc:`NoSuchDeviceError`, if no device was found for ``sys_path``.
+
+        .. versionchanged:: 0.4
+           Raise :exc:`NoSuchDeviceError` instead of returning ``None``, if
+           no device was found for ``sys_path``
         """
         if not isinstance(context, Context):
             raise TypeError('Invalid context object')
         device = libudev.udev_device_new_from_syspath(
             context._context, _assert_bytes(sys_path))
         if not device:
-            return None
+            raise NoSuchDeviceError(sys_path)
         return cls(context, device)
 
     def __init__(self, context, _device):